Simplify Scanner.Scan method, remove done channel
The output channel can be closed to signify the end of the scan when
the output is consumed concurrently.

The done chaannel was usuful to wait until the end of the scan before
consuming its output.

I think that requiring two parameters with the same role in two
different use cases is likely a sign that there should really be two
distinct methods.